# swagger.json
{
  "definitions": {},
  "info": {
    "title": "Schedule Appointment API",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "description": "API for managing schedules and appointments"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "/",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/schedules": {
      "get": {
        "summary": "/schedules",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      },
      "post": {
        "summary": "/schedules",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/schedules/{id}": {
      "get": {
        "summary": "/schedules/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "put": {
        "summary": "/schedules/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "delete": {
        "summary": "/schedules/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/schedules/users/{userId}": {
      "get": {
        "summary": "/schedules/users/{userId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "delete": {
        "summary": "/schedules/users/{userId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/schedules/sections/{sectionId}": {
      "get": {
        "summary": "/schedules/sections/{sectionId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/appointments": {
      "get": {
        "summary": "/appointments",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "304": {}
        },
        "tags": []
      },
      "post": {
        "summary": "/appointments",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {},
        "tags": []
      }
    },
    "/appointments/{id}": {
      "get": {
        "summary": "/appointments/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "put": {
        "summary": "/appointments/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "delete": {
        "summary": "/appointments/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/appointments/users/{userId}": {
      "get": {
        "summary": "/appointments/users/{userId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      },
      "delete": {
        "summary": "/appointments/users/{userId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    },
    "/appointments/schedules/{scheduleId}": {
      "get": {
        "summary": "/appointments/schedules/{scheduleId}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "scheduleId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {},
        "tags": []
      }
    }
  },
  "swagger": "2.0",
  "tags": []
}